+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- literate: report missing closing of codeblock
+
 ## 0.8.0
 
 - literate programming, first backend: [tr-notes](https://tr-notes.srht.site/)